requirements.txt
python-telegram-bot==10.1.0
beautifulsoup4==4.6.0
lxml
